feat: search for the task before finish or exclude

+++ app/backend/src/services/tasks.service.js
@@ -36,6 +36,10 @@ const create = async (description, userId) => {
 };
 
 const finish = async (id) => {
+  const foundTask = await Task.findOne({ where: id });
+
+  if (!foundTask) return { message: 'Tarefa não encontrada' };
+
   await Task.update(
     { status: 'Concluído' },
     { where: { id } },
@@ -45,6 +49,10 @@ const finish = async (id) => {
 };
 
 const exclude = async (id) => {
+  const foundTask = await Task.findOne({ where: id });
+
+  if (!foundTask) return { message: 'Tarefa não encontrada' };
+
   await Task.destroy({ where: { id } });
 };
